chore(map): move to `lib`.
