+++ src/map/index.ts
@@ -1,4 +1,1 @@
-export { CoreMap } from './core-map.abstract';
-export { DataMap } from './data-map.class';
-export { FactoryMap } from './factory-map.class';
-export { WeakDataMap } from './weak-data-map.class';
+export * from './lib';
