+++ src/map/lib/index.ts
@@ -1,0 +1,4 @@
+export { CoreMap } from './core-map.abstract';
+export { DataMap } from './data-map.class';
+export { FactoryMap } from './factory-map.class';
+export { WeakDataMap } from './weak-data-map.class';
